# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Agronom Bot Internal API
  version: "1.10.0"
  description: |
    Telegram-бот «Карманный агроном» — API v1.10.0.
    Fixes: +signature, +UNAUTHORIZED, +monthly limits, +X-API-Ver usage, +multipart required, clarified dosage units, +X-User-ID header, +ask_expert endpoint, +protocol fields category/status/waiting_days, +assistant chat/confirm_plan.
  contact:
    name: Ivan Gromov
    url: https://github.com/agronom-bot
    email: i.gromov@agrostore.ru
servers:
  - url: https://api.agronom.local
    description: Production
  - url: https://staging.agronom.local
    description: Staging
security:
  - ApiKeyAuth: []
    RequestSignAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    RequestSignAuth:
      type: apiKey
      in: header
      name: X-Req-Sign
    HmacAuth:
      type: apiKey
      in: header
      name: X-Sign
  parameters:
    ApiVersionHeader:
      name: X-API-Ver
      in: header
      required: true
      schema:
        type: string
        enum: [v1]
    UserIdHeader:
      name: X-User-ID
      in: header
      required: true
      schema:
        type: integer
    RequestTimestampHeader:
      name: X-Req-Ts
      in: header
      required: true
      schema:
        type: integer
        description: Unix timestamp (seconds)
    RequestNonceHeader:
      name: X-Req-Nonce
      in: header
      required: true
      schema:
        type: string
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            [NO_LEAF, LIMIT_EXCEEDED, GPT_TIMEOUT, BAD_REQUEST, UNAUTHORIZED, UPGRADE_REQUIRED, TOO_MANY_REQUESTS, SERVICE_UNAVAILABLE, FORBIDDEN]
        message:
          type: string
...
    PlanCreateRequest:
      type: object
      required: [object_id, plan_payload]
      properties:
        object_id:
          type: integer
        case_id:
          type: integer
          nullable: true
        source:
          type: string
          example: assistant
        plan_payload:
          type: object
          additionalProperties: true
    PlanCreateResponse:
      type: object
      properties:
        plan_id:
          type: integer
          nullable: true
        stages:
          type: array
          items:
            $ref: '#/components/schemas/PlanCreateStage'
        errors:
          type: array
          items:
            type: string
...
paths:
...
  /v1/plans:
    post:
      summary: Create plan from plan_payload
      operationId: createPlan
      description: "Создаёт план по plan_payload (assistant/master)."
      tags: [plans]
      security:
        - ApiKeyAuth: []
          RequestSignAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/RequestTimestampHeader'
        - $ref: '#/components/parameters/RequestNonceHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanCreateRequest'
      responses:
        '200':
          description: Plan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanCreateResponse'
        '400':
          description: Bad payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/plans/{plan_id}/pdf:
    get:
      summary: Export plan to PDF
      operationId: exportPlanPdf
      description: "Экспортирует план в PDF-заметку."
      tags: [plans]
      security:
        - ApiKeyAuth: []
          RequestSignAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/RequestTimestampHeader'
        - $ref: '#/components/parameters/RequestNonceHeader'
        - name: plan_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/plans/{plan_id}/events:
    post:
      summary: Create event/reminder for plan
      operationId: createPlanEvent
      description: "Создаёт событие плана и напоминание (при необходимости)."
      tags: [plans]
      security:
        - ApiKeyAuth: []
          RequestSignAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/RequestTimestampHeader'
        - $ref: '#/components/parameters/RequestNonceHeader'
        - name: plan_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreateRequest'
      responses:
        '200':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventCreateResponse'
        '400':
          description: Bad payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/plans/{plan_id}/autoplan:
    post:
      summary: Enqueue autoplan run
      operationId: enqueueAutoplan
      description: "Создаёт запись автоплана для поиска окна обработок."
      tags: [plans]
      security:
        - ApiKeyAuth: []
          RequestSignAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/RequestTimestampHeader'
        - $ref: '#/components/parameters/RequestNonceHeader'
        - name: plan_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutoplanRequest'
      responses:
        '202':
          description: Autoplan enqueued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoplanResponse'
        '400':
          description: Bad payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/plans/sessions:
    post:
      summary: Create or refresh plan wizard session state
      operationId: upsertPlanSession
      description: "Сохраняет шаг мастера плана или обновляет существующую сессию."
      tags: [plans]
      security:
        - ApiKeyAuth: []
          RequestSignAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/RequestTimestampHeader'
        - $ref: '#/components/parameters/RequestNonceHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanSessionUpsertRequest'
      responses:
        '200':
          description: Session stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanSessionRecord'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Fetch the latest or token-specific plan wizard session
      operationId: getPlanSession
      description: "Возвращает последнюю сессию или конкретный токен мастера планов."
      tags: [plans]
      security:
        - ApiKeyAuth: []
          RequestSignAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/RequestTimestampHeader'
        - $ref: '#/components/parameters/RequestNonceHeader'
        - name: token
          in: query
          required: false
          schema:
            type: string
            description: "Если не указан, возвращается последняя сессия пользователя."
        - name: include_expired
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: plan_id
          in: query
          required: false
          schema:
            type: integer
            description: "Вернуть сессию, связанную с планом (взаимоисключается с token)."
      responses:
        '200':
          description: Session returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanSessionRecord'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Session expired (unless include_expired=true)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete plan wizard session(s)
      operationId: deletePlanSession
      description: "Удаляет конкретную сессию по токену/плану или все сессии пользователя."
      tags: [plans]
      security:
        - ApiKeyAuth: []
          RequestSignAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/RequestTimestampHeader'
        - $ref: '#/components/parameters/RequestNonceHeader'
        - name: token
          in: query
          required: false
          schema:
            type: string
            description: "Если не указан, удаляются все сессии пользователя."
        - name: plan_id
          in: query
          required: false
          schema:
            type: integer
            description: "Удалить сессии, связанные с планом."
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/plans/sessions/{session_id}:
    patch:
      summary: Update plan wizard session fields
      operationId: patchPlanSession
      description: "Частично обновляет состояние конкретной сессии."
      tags: [plans]
      security:
        - ApiKeyAuth: []
          RequestSignAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/RequestTimestampHeader'
        - $ref: '#/components/parameters/RequestNonceHeader'
        - name: session_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanSessionPatchRequest'
      responses:
        '200':
          description: Updated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanSessionRecord'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...